xParrot is now an optionally-interactive CLI!
Note that I took away the option to manually specify the API key because I couldn't figure out a way to get something like this to work:

```
xparrot -i -k keyXXXxxxXXX
```

The key always came through in the inital docopt-parsed args, but then it wouldn't be available in the tasks function. No idea why, but this is fine to be getting on with. Just set AIRTABLE_API_KEY in your environment, per the revised docs.